--- base-html/invitation-deloge.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="fr">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <link
-      rel="stylesheet"
-      href="https://cdn.rawgit.com/mblode/marx/master/css/marx.min.css"
-    />
-    <title>Document</title>
-  </head>
-  <body>
-    <form action="" onsubmit="submitForm(event)" id="form">
-      <h3>Créer un compte "délogé"</h3>
-      <label for="nom">Nom</label> <input type="text" name="nom" id="nom" />
-      <label for="prenom">Prénom</label>
-      <input type="text" name="prenom" id="prenom" />
-      <legend>
-        Merci d'indiquer au minimum soit l'email, soit le numéro de téléphone :
-      </legend>
-      <label for="email">Email</label>
-      <input type="email" name="email" id="email" />
-      <label for="tel">Téléphone</label>
-      <input type="tel" name="tel" id="tel" />
-      <button type="submit">Générer une invitation</button>
-    </form>
-    <script src="form.js"></script>
-  </body>
-</html>
