--- base-html/parametres-benevole.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html lang="fr">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <link
-      rel="stylesheet"
-      href="https://cdn.rawgit.com/mblode/marx/master/css/marx.min.css"
-    />
-    <title>Document</title>
-  </head>
-  <body>
-    <form action="" onsubmit="submitForm(event)" id="form">
-      <h3>Mes paramètres</h3>
-      <fieldset>
-        <legend>J'offre :</legend>
-        <input type="checkbox" id="restauration" name="restauration" />
-        <label for="restauration">Restauration</label>
-        <input type="checkbox" id="blanchisserie" name="blanchisserie" />
-        <label for="blanchisserie">Blanchisserie</label>
-        <input type="checkbox" id="mobilité" name="mobilité" />
-        <label for="mobilité">Mobilité</label>
-        <input type="checkbox" id="loisirs" name="loisirs" />
-        <label for="loisirs">Loisirs</label>
-        <input type="checkbox" id="Don" name="Don" />
-        <label for="Don">Don</label>
-        <input type="checkbox" id="autre" name="autre" />
-        <label for="autre">Autre(s) service(s)</label>
-      </fieldset>
-    </form>
-    <script src="form.js"></script>
-  </body>
-</html>
