+++ package.json
@@ -1,0 +1,10 @@
+{
+  "name": "aouf",
+  "version": "0.1.0",
+  "private": true,
+  "workspaces": {
+    "packages": [
+      "packages/*"
+    ]
+  }
+}
