+++ packages/api/package.json
@@ -1,20 +1,17 @@
 {
-  "name": "hackathon-aouf",
+  "name": "@aouf/api",
   "version": "0.1.0",
   "description": "",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon server",
-    "lint": "eslint ./",
-    "create-database": "node ./scripts/create-database",
-    "hash-password": "node ./scripts/hash-password"
+    "start": "nodemon src",
+    "lint": "eslint ./"
   },
   "author": "",
   "license": "MIT",
   "dependencies": {
     "bcryptjs": "^2.4.3",
-    "body-parser": "^1.18.3",
     "cors": "^2.8.5",
     "dotenv": "^6.2.0",
     "express": "^4.16.4",
@@ -23,14 +20,12 @@
     "jsonwebtoken": "^8.4.0",
     "knex": "^0.16.3",
     "morgan": "^1.9.1",
-    "mysql2": "^1.6.4",
-    "nunjucks": "^3.1.7"
+    "mysql2": "^1.6.4"
   },
   "devDependencies": {
     "eslint": "^5.12.1",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-config-prettier": "^3.6.0",
-    "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-prettier": "^3.0.1",
     "nodemon": "^1.18.9",
     "prettier": "^1.15.3"
