+++ packages/api/src/errors.js
