--- server/api/dislodged/README.md
@@ -1,9 +1,0 @@
-# [Aouf](../../README.md) > [API](../README.md) > Dislodged
-
-## signup
-
-Crée un compte délogé à partir d'un token d'inscription.
-
-## signupToken
-
-Crée un token d'inscription pour un délogé.
