--- server/api/signout.js
@@ -1,6 +1,0 @@
-const signout = async (req, res) => {
-  req.session.destroy();
-  res.json(true);
-};
-
-module.exports = signout;
