--- server/api/volunteer/README.md
@@ -1,5 +1,0 @@
-# [Aouf](../../README.md) > [API](../README.md) > Volunteer
-
-## infos
-
-Renvoie les information d'un bénévole.
