--- server/controllers/README.md
@@ -1,8 +1,0 @@
-# [Aouf](../README.md) > Controllers
-
-Un controller est une fonction qui réalise une ou des opérations relative(s) à la requête et/ou la réponse.  
-Le controller renvoie la réponse à la fin de son traitement.
-
-> **request** -> middleware1 -> middleware2 -> … -> controller -> **response**
-
-Au moment de la rédaction de ce README, tous les controllers de ce dossier retourne une page HTML, les controllers d'api ayant été disposé dans un dossier à part ([/api](../api/README.md))
