--- server/controllers/helloworld.js
@@ -1,3 +1,0 @@
-module.exports = (req, res) => {
-  res.render('helloworld.html');
-};
