--- server/controllers/inscription-benevole.js
@@ -1,3 +1,0 @@
-module.exports = (req, res) => {
-  res.render('inscription-benevole.html');
-};
