--- server/controllers/invitation-deloge.js
@@ -1,3 +1,0 @@
-module.exports = (req, res) => {
-  res.render('invitation-deloge.html');
-};
