--- server/public/humans.txt
@@ -1,3 +1,0 @@
-/* SITE */                      
-  Last update:  2019/01/19             
-  Standards: HTML5, CSS3
