--- server/scripts/hash-password.js
@@ -1,5 +1,0 @@
-const bcrypt = require('bcryptjs');
-const [, , password] = process.argv;
-
-// eslint-disable-next-line no-console
-console.log(bcrypt.hashSync(password, 10));
