--- server/views/document.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="{{ lang|default('fr')}}">
-  <head>
-    <meta charset="utf-8" />
-    <title>Aouf</title>
-    <link type="text/plain" rel="author" href="/assets/humans.txt" />
-    <link
-      rel="stylesheet"
-      href="https://cdn.rawgit.com/mblode/marx/master/css/marx.min.css"
-    />
-    <style>
-      body {
-        margin: 0 auto;
-        max-width: 600px;
-      }
-    </style>
-    {% block head %}{% endblock %}
-  </head>
-  <body>
-    {% block body %}{% endblock %}
-  </body>
-</html>
