--- server/views/helloworld.html
@@ -1,1 +1,0 @@
-{% extends "document.html" %} {% block body %} Hello world! {% endblock %}
